Made the information about styling a bit clearer.
I spent an hour debugging the "Interface 'Maatwebsite\Excel\Concerns\WithStyles' not found" message. According to the documentation, version ^3.1 must allow me to use `WithStyles` Concern, but my laravel version is not compatible with the latest (v3.1.26 at that moment) version of laravel-excel. I've added the version number that allows the use of this Concern to the documentation, so less people bang their heads :)