Added documentation to further clarify how to download the Excel file.  Since this function was named 'download' as well, it is confusing to me if this was to be part of InvoiceExport, or the Sheet class.  Finally, I understood this is just example code that could be implemented anywhere.
Renaming it to downloadInvoices, and adding extra documentation clarifies that it is code that could be implemented in a controller etc.

Also, on first run, I never included WithTitle, and had to cycle back to the documentation since it was not explicitly mentioned that WithTitle was needed.

Instead, I only implemented the title() method, and didn't know why it wasn't working.

By adding an explicit note, that WithTitle concern is needed, the reader may be saved time.